fix: improve memory performances for VDA model
This commit allows to use the custom version of VideoDepthAnything:
OptimizedVideoDepthAnything.

This modification slightly improves memory performances by splitting the
process method into 2 steps (inference and alignment). This also
improves the way variables are assigned and try to avoid numpy array
duplication as much as possible.

While chunk processing would be a more robust solution, implementing it
is challenging due to the sliding window mechanism used during
processing and the alignment step, which requires all generated depths.
A more comprehensive refactor could address this, but it would require
significantly more effort and time.

As a result, chunk processing is set to disabled in Meshroom when VDA is
selected for now.